Jsonlogic roundtrip fix (#1092)
* First solution

This uses simple operator checks of jsonlogic

* Jsonlogic roundtrip should work as intended now

For operators like select_not_any_in where the ! is part of it we now always assume that the "!" is part of the operator and translate the jsonlogic combination straight to operator. We no longer use any reversing to try and achieve this.

Jsonlogic import has now new functions for matching these jsonlogic combinations to operator. Its use could be expanded further to reduce the complexity of the code and to possibly better enable use of custom jsonlogic defintions for operators.

Export has less than ideal solution but it works it seems.
Proper tests for the roundtrip are still needed. Also some of the tests dont work now that import works properly. Also i think i found identical issue with spel as existed with jsonlogic when looking at tests.

* New special case check during export and tests

* Removed console logs from develpoment

* lint fixes

* remove unused _jsonLogicIsRevArgs

* Fix for func, added test

* tests for reverseOperatorsForNot fail

* Tests fixed and improved

* Improved tests and changed !group negation handling

* fix for reverseOperatorsForNot == true

This reverts commit 69aee11c4ff6ce7c1d8d6f8f0bfa39ff54d3d4af.

* chlog

---------

Co-authored-by: ukrbublik <ukrbublik@gmail.com>